requirements.txt
matplotlib==3.8.1
numpy==1.26.1
onnx==1.15.0
opencv-python==4.8.1.78
pillow==9.4.0
pandas==2.1.2
pyqt5==5.15.2
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchsummary==1.5.1
torchvision==0.13.0+cu113
tqdm==4.66.1
yolov5==7.0.13